removing zeslot configuration logic from zepreferences service
